refactor: 최근 검색어 리스트 조회 API DDD + hexagonal로 리팩토링
