--- src/main/java/com/joojoo/api/search/application/SearchService.java
@@ -1,8 +1,0 @@
-package com.joojoo.api.search.application;
-
-import com.joojoo.api.search.presentation.dto.response.RecentSearchListResponse;
-
-public interface SearchService {
-
-    RecentSearchListResponse recordSearchList(Long userId);
-}
