+++ src/main/java/com/joojoo/api/search/application/port/in/GetSearchUseCase.java
@@ -1,7 +1,10 @@
 package com.joojoo.api.search.application.port.in;
 
+import com.joojoo.api.search.presentation.dto.response.RecentSearchListResponse;
 import com.joojoo.api.search.presentation.dto.response.TagHistoryResponseDto;
 
 public interface GetSearchUseCase {
     TagHistoryResponseDto searchTags(Long userId, String query);
+
+    RecentSearchListResponse recordSearchList(Long userId);
 }
