+++ src/main/java/com/joojoo/api/search/application/service/query/GetSearchService.java
@@ -1,9 +1,13 @@
 package com.joojoo.api.search.application.service.query;
 
 import com.joojoo.api.blockTag.infrastructure.redis.BlockTagRedisRepository;
+import com.joojoo.api.common.domain.enums.SearchTarget;
 import com.joojoo.api.common.search.SearchRangeFactory;
 import com.joojoo.api.common.search.range.SearchRange;
 import com.joojoo.api.search.application.port.in.GetSearchUseCase;
+import com.joojoo.api.search.domain.repository.SearchRepository;
+import com.joojoo.api.search.presentation.dto.request.RecentSearchDto;
+import com.joojoo.api.search.presentation.dto.response.RecentSearchListResponse;
 import com.joojoo.api.search.presentation.dto.response.TagHistoryResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -11,6 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 
 import java.util.Collections;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -18,7 +23,7 @@ import java.util.Collections;
 public class GetSearchService implements GetSearchUseCase {
 
     private final SearchRangeFactory searchRangeFactory;
-
+    private final SearchRepository searchRepository;
     private final BlockTagRedisRepository blockTagRedisRepository;
 
     @Override
@@ -27,6 +32,21 @@ public class GetSearchService implements GetSearchUseCase {
 
         SearchRange range = searchRangeFactory.createSearchTagKeyRange(query, userId);
         return TagHistoryResponseDto.from(blockTagRedisRepository.searchTagQuery(range));
+    }
+
+    @Override
+    public RecentSearchListResponse recordSearchList(Long userId) {
+        return RecentSearchListResponse.builder()
+                .tickers(fetchRecent(userId, SearchTarget.TICKER))
+                .tags(fetchRecent(userId, SearchTarget.TAG))
+                .build();
+    }
+
+    private List<RecentSearchDto> fetchRecent(Long userId, SearchTarget type) {
+        return searchRepository.findRecentByTargetType(userId, type, 10)
+                .stream()
+                .map(RecentSearchDto::of)
+                .toList();
     }
 
 }
